requirements.txt
APScheduler==3.6.3
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-Login==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paho-mqtt==1.5.1
pyasn1==0.4.8
pymongo==3.11.1
pytz==2020.4
rsa==4.6
six==1.15.0
tzlocal==2.1
Werkzeug==1.0.1
